Replaced src with a cmake version and initialized folder for release
