--- gui/moc_mainwindow.cpp
@@ -1,201 +1,0 @@
-/****************************************************************************
-** Meta object code from reading C++ file 'mainwindow.h'
-**
-** Created by: The Qt Meta Object Compiler version 68 (Qt 6.7.2)
-**
-** WARNING! All changes made in this file will be lost!
-*****************************************************************************/
-
-#include "../../../mainwindow.h"
-#include <QtCore/qmetatype.h>
-
-#include <QtCore/qtmochelpers.h>
-
-#include <memory>
-
-
-#include <QtCore/qxptype_traits.h>
-#if !defined(Q_MOC_OUTPUT_REVISION)
-#error "The header file 'mainwindow.h' doesn't include <QObject>."
-#elif Q_MOC_OUTPUT_REVISION != 68
-#error "This file was generated using the moc from 6.7.2. It"
-#error "cannot be used with the include files from this version of Qt."
-#error "(The moc has changed too much.)"
-#endif
-
-#ifndef Q_CONSTINIT
-#define Q_CONSTINIT
-#endif
-
-QT_WARNING_PUSH
-QT_WARNING_DISABLE_DEPRECATED
-QT_WARNING_DISABLE_GCC("-Wuseless-cast")
-namespace {
-
-#ifdef QT_MOC_HAS_STRINGDATA
-struct qt_meta_stringdata_CLASSMainWindowENDCLASS_t {};
-constexpr auto qt_meta_stringdata_CLASSMainWindowENDCLASS = QtMocHelpers::stringData(
-    "MainWindow",
-    "on_lineEdit_textChanged",
-    "",
-    "arg1",
-    "on_listView_activated",
-    "QModelIndex",
-    "index",
-    "on_listView_clicked",
-    "on_lineEdit_editingFinished",
-    "on_pushButton_clicked",
-    "on_autoCompleteView_activated",
-    "on_autoCompleteView_clicked",
-    "on_inputBox2_textChanged",
-    "on_pushButton_2_clicked",
-    "on_resultWidget_itemSelectionChanged",
-    "select",
-    "qsizetype"
-);
-#else  // !QT_MOC_HAS_STRINGDATA
-#error "qtmochelpers.h not found or too old."
-#endif // !QT_MOC_HAS_STRINGDATA
-} // unnamed namespace
-
-Q_CONSTINIT static const uint qt_meta_data_CLASSMainWindowENDCLASS[] = {
-
- // content:
-      12,       // revision
-       0,       // classname
-       0,    0, // classinfo
-      12,   14, // methods
-       0,    0, // properties
-       0,    0, // enums/sets
-       0,    0, // constructors
-       0,       // flags
-       0,       // signalCount
-
- // slots: name, argc, parameters, tag, flags, initial metatype offsets
-       1,    1,   86,    2, 0x08,    1 /* Private */,
-       4,    1,   89,    2, 0x08,    3 /* Private */,
-       7,    1,   92,    2, 0x08,    5 /* Private */,
-       8,    0,   95,    2, 0x08,    7 /* Private */,
-       9,    0,   96,    2, 0x08,    8 /* Private */,
-      10,    1,   97,    2, 0x08,    9 /* Private */,
-      11,    1,  100,    2, 0x08,   11 /* Private */,
-      12,    1,  103,    2, 0x08,   13 /* Private */,
-      13,    0,  106,    2, 0x08,   15 /* Private */,
-      14,    0,  107,    2, 0x08,   16 /* Private */,
-      15,    1,  108,    2, 0x08,   17 /* Private */,
-      15,    0,  111,    2, 0x28,   19 /* Private | MethodCloned */,
-
- // slots: parameters
-    QMetaType::Void, QMetaType::QString,    3,
-    QMetaType::Void, 0x80000000 | 5,    6,
-    QMetaType::Void, 0x80000000 | 5,    6,
-    QMetaType::Void,
-    QMetaType::Void,
-    QMetaType::Void, 0x80000000 | 5,    6,
-    QMetaType::Void, 0x80000000 | 5,    6,
-    QMetaType::Void, QMetaType::QString,    3,
-    QMetaType::Void,
-    QMetaType::Void,
-    QMetaType::Void, 0x80000000 | 16,    6,
-    QMetaType::Void,
-
-       0        // eod
-};
-
-Q_CONSTINIT const QMetaObject MainWindow::staticMetaObject = { {
-    QMetaObject::SuperData::link<QMainWindow::staticMetaObject>(),
-    qt_meta_stringdata_CLASSMainWindowENDCLASS.offsetsAndSizes,
-    qt_meta_data_CLASSMainWindowENDCLASS,
-    qt_static_metacall,
-    nullptr,
-    qt_incomplete_metaTypeArray<qt_meta_stringdata_CLASSMainWindowENDCLASS_t,
-        // Q_OBJECT / Q_GADGET
-        QtPrivate::TypeAndForceComplete<MainWindow, std::true_type>,
-        // method 'on_lineEdit_textChanged'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        QtPrivate::TypeAndForceComplete<const QString &, std::false_type>,
-        // method 'on_listView_activated'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        QtPrivate::TypeAndForceComplete<const QModelIndex &, std::false_type>,
-        // method 'on_listView_clicked'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        QtPrivate::TypeAndForceComplete<const QModelIndex &, std::false_type>,
-        // method 'on_lineEdit_editingFinished'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        // method 'on_pushButton_clicked'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        // method 'on_autoCompleteView_activated'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        QtPrivate::TypeAndForceComplete<const QModelIndex &, std::false_type>,
-        // method 'on_autoCompleteView_clicked'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        QtPrivate::TypeAndForceComplete<const QModelIndex &, std::false_type>,
-        // method 'on_inputBox2_textChanged'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        QtPrivate::TypeAndForceComplete<const QString &, std::false_type>,
-        // method 'on_pushButton_2_clicked'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        // method 'on_resultWidget_itemSelectionChanged'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        // method 'select'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>,
-        QtPrivate::TypeAndForceComplete<qsizetype, std::false_type>,
-        // method 'select'
-        QtPrivate::TypeAndForceComplete<void, std::false_type>
-    >,
-    nullptr
-} };
-
-void MainWindow::qt_static_metacall(QObject *_o, QMetaObject::Call _c, int _id, void **_a)
-{
-    if (_c == QMetaObject::InvokeMetaMethod) {
-        auto *_t = static_cast<MainWindow *>(_o);
-        (void)_t;
-        switch (_id) {
-        case 0: _t->on_lineEdit_textChanged((*reinterpret_cast< std::add_pointer_t<QString>>(_a[1]))); break;
-        case 1: _t->on_listView_activated((*reinterpret_cast< std::add_pointer_t<QModelIndex>>(_a[1]))); break;
-        case 2: _t->on_listView_clicked((*reinterpret_cast< std::add_pointer_t<QModelIndex>>(_a[1]))); break;
-        case 3: _t->on_lineEdit_editingFinished(); break;
-        case 4: _t->on_pushButton_clicked(); break;
-        case 5: _t->on_autoCompleteView_activated((*reinterpret_cast< std::add_pointer_t<QModelIndex>>(_a[1]))); break;
-        case 6: _t->on_autoCompleteView_clicked((*reinterpret_cast< std::add_pointer_t<QModelIndex>>(_a[1]))); break;
-        case 7: _t->on_inputBox2_textChanged((*reinterpret_cast< std::add_pointer_t<QString>>(_a[1]))); break;
-        case 8: _t->on_pushButton_2_clicked(); break;
-        case 9: _t->on_resultWidget_itemSelectionChanged(); break;
-        case 10: _t->select((*reinterpret_cast< std::add_pointer_t<qsizetype>>(_a[1]))); break;
-        case 11: _t->select(); break;
-        default: ;
-        }
-    }
-}
-
-const QMetaObject *MainWindow::metaObject() const
-{
-    return QObject::d_ptr->metaObject ? QObject::d_ptr->dynamicMetaObject() : &staticMetaObject;
-}
-
-void *MainWindow::qt_metacast(const char *_clname)
-{
-    if (!_clname) return nullptr;
-    if (!strcmp(_clname, qt_meta_stringdata_CLASSMainWindowENDCLASS.stringdata0))
-        return static_cast<void*>(this);
-    return QMainWindow::qt_metacast(_clname);
-}
-
-int MainWindow::qt_metacall(QMetaObject::Call _c, int _id, void **_a)
-{
-    _id = QMainWindow::qt_metacall(_c, _id, _a);
-    if (_id < 0)
-        return _id;
-    if (_c == QMetaObject::InvokeMetaMethod) {
-        if (_id < 12)
-            qt_static_metacall(this, _c, _id, _a);
-        _id -= 12;
-    } else if (_c == QMetaObject::RegisterMethodArgumentMetaType) {
-        if (_id < 12)
-            *reinterpret_cast<QMetaType *>(_a[0]) = QMetaType();
-        _id -= 12;
-    }
-    return _id;
-}
-QT_WARNING_POP
