--- src/lists/categories.txt
@@ -1,34 +1,0 @@
-nan
-Toothpaste
-Tooth Whitening
-Hand Sanitizer
-Liquid Hand Soap
-Antiperspirant/Deodorant
-Baby Toothpaste
-Body Wash/Cleanser
-Mouthwash
-Lip Balm
-Muscle/Joint Soreness
-Baby Soap
-Body Oil
-Foot Moisturizer
-Moisturizer
-Blush
-Exfoliant/Scrub
-Baby Lotion
-Around Eye Cream
-Bath Oil/Salts/Soak
-Cuticle Treatment
-Facial Moisturizer/Treatment
-Foot Odor Control
-Skin Fading/Lightener
-Bar Soap
-Eye Shadow
-Hand Cream
-lip balms with SPF
-recreational sunscreens
-Baby Wipes
-Body Firming Lotion
-Anti Aging
-Facial Cleanser
-Concealer
