--- src/scraper/save.txt
@@ -1,2 +1,0 @@
-,Product,Category,Ingredients
-0,,,
