--- src/srcQT/gui/gui.pro
@@ -1,35 +1,0 @@
-QT       += core gui
-
-greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
-
-CONFIG += c++17
-
-# You can make your code fail to compile if it uses deprecated APIs.
-# In order to do so, uncomment the following line.
-#DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
-
-SOURCES += \
-    embed.cpp \
-    main.cpp \
-    mainwindow.cpp
-    embed.cpp
-
-HEADERS += \
-    embed.h \
-    mainwindow.h
-    embed.h
-
-FORMS += \
-    mainwindow.ui
-
-# Default rules for deployment.
-qnx: target.path = /tmp/$${TARGET}/bin
-else: unix:!android: target.path = /opt/$${TARGET}/bin
-!isEmpty(target.path): INSTALLS += target
-
-INCLUDEPATH += C:/Python312/include
-LIBS += -LC:/Python312/libs -lpython312_d
-
-DISTFILES +=
-
-
