--- src/srcQT/readme.txt
@@ -1,2 +1,0 @@
-THIS FOLDER IS FOR BUILDING QT FROM SOURCE.
-FOR WINDOWS 11 YOU CAN JUST USE THE PREBUILT BINARIES.
